Bluetooth: host: directed advertising support
Added implementation to the directed advertising API in the Connection
Management module. Introduced a new connection state for this type of
advertising. The new state is symmetric to the connection state used for
scanning.

Added new advertising options that can be used to trigger low and high
duty directed advertising. Extended the Advertising Parameters
descriptor, so now it includes peer address. Added macros for default
values of Advertising Parameters, which are used to trigger directed
advertising.

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>